+++ modules/client/src/tokenVoting/internal/client/estimation.ts
@@ -8,7 +8,11 @@ import {
 } from "../../../client-common";
 import { ITokenVotingClientEstimation } from "../interfaces";
 import { toUtf8Bytes } from "@ethersproject/strings";
-import { boolArrayToBitmap, decodeProposalId } from "@aragon/sdk-common";
+import {
+  boolArrayToBitmap,
+  decodeProposalId,
+  SizeMismatchError,
+} from "@aragon/sdk-common";
 import { DelegateTokensParams } from "../../types";
 import { ClientCore, GasFeeEstimation } from "@aragon/sdk-client-common";
 /**
@@ -37,9 +41,7 @@ export class TokenVotingClientEstimation extends ClientCore
       params.failSafeActions?.length &&
       params.failSafeActions.length !== params.actions?.length
     ) {
-      throw new Error(
-        "Size mismatch: actions and failSafeActions should match",
-      );
+      throw new SizeMismatchError();
     }
     const allowFailureMap = boolArrayToBitmap(params.failSafeActions);